Change architecture of the bootloader
This commit changes the architecture of the bootloader. Previous to that
the bootloader was developed on the assumption that it is possible to
replace factory programmed bootloader located in System Memory. It is
not true. Because of that limitation and to circumvent it the
following layout of the Application Flash was implemented:
 - [0x08000000 - 0x08004800] Bootloader
   This memory is reserved for STFUBoot and it's code
 - [0x08004800 - 0x08004A00] Firmware Info
   This area contains 512-bytes header that contains information about
   the user firmware image, such as size and checksum
 - [0x08004A00 - end of Flash] Application
   This are where the application firmware is written

This commit also introduces user space tool 'stfub-prefix' which can
be used to add necessary for firmware integrity checking prefixes to
the binary images.

As of now the implementation is in the semi-working condition with FW
integrity check being omitted by the bootloader.