Add more robust exception handling.
This commit changes the behaviour of the bootloader in the event of any
exception. Now bootolader will jump to factory built-in bootloader in
the event of the exception happening in the bootloader code(Memory
Managment exception, Hard Fault, etc.)

It also introduces "flash" target to the Makefile that allows update
the bootloader using the bootloader itself, providing that the initial
FW was flashed with ST DFU implementation.