Implemented flashing, CRC, then jumping to app
The python utility was modified to work with the new CRC code. It also
creates a new file rather than modify the original .bin.

The CRC code with jumping to the application was implemented. There are
sometimes errors with the flashing and the unit needs to be reset and
reflashed. It however does not get bricked.